Major code reorganisation
As part of the transition from subversion to git it was time to get
rid of the trunk directories.
The code layout was changed a bit as well, all server relevant code,
including deployment, is now found under core/.
This commit will break many things (most notably the package build)
because the implicit assumptions about the directory structure
have not yet been corrected.